LRSD-1334 Rename AttacmenUploaderMessage
